Updated controller
